# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Serviço Backend",
        "description": "API RESTful para gerenciamento de serviços comunitários, incluindo pessoas, perfis, missões, notícias e doações.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/perfis": {
            "get": {
                "tags": [
                    "Perfis"
                ],
                "summary": "Lista todos os perfis",
                "operationId": "c81e2826c91f75ac01a11c02c719ef50",
                "responses": {
                    "200": {
                        "description": "Lista de perfis retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Perfis"
                ],
                "summary": "Cria um novo perfil",
                "operationId": "f9ae9bf070c1fd189bb553668648b676",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "ADMIN"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Administrador do sistema"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Perfil criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/api/perfis/{id}": {
            "get": {
                "tags": [
                    "Perfis"
                ],
                "summary": "Busca perfil por ID",
                "operationId": "140b42070528553a5a5bb773ddf468e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do perfil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perfil encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Perfil não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Perfis"
                ],
                "summary": "Atualiza um perfil",
                "operationId": "562461cea21af54561cccd85425634d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Perfil não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Perfis"
                ],
                "summary": "Exclui um perfil",
                "operationId": "de3cfadb8c341949bd752e3381ceabf0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Perfil excluído com sucesso"
                    },
                    "404": {
                        "description": "Perfil não encontrado"
                    },
                    "409": {
                        "description": "Não é possível excluir, existem pessoas vinculadas"
                    }
                }
            }
        },
...
    },
...
}